% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_deskriptive_statistik.R
\name{absolute_density_grouped_data}
\alias{absolute_density_grouped_data}
\title{Compute the absolute density for grouped data}
\usage{
absolute_density_grouped_data(x, levels)
}
\arguments{
\item{x}{A numeric vector.}

\item{levels}{Levels of a factor, usually created by \code{\link[base]{cut}}.}
}
\description{
Compute the absolute density for grouped data
}
