% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{actionItem}
\alias{actionItem}
\title{Action item}
\usage{
actionItem(inputId, label, icon = NULL, width = NULL)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link - usually a text label, but
you could also use any other HTML, like an image.}

\item{width}{The width of the input, e.g. \code{'400px'}, or \code{'100\%'};
see \code{\link[shiny]{validateCssUnit}}.}

\item{An}{optional \code{\link[shiny]{icon}} to appear on the button.}
}
\description{
Create a \code{\link[shinydashboard:sidebarMenu]{menuItem}}-like
\code{\link[shiny]{actionButton}} with an enclosing div with class
"div-btn-sidebar".
}
