% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{actionSubItem}
\alias{actionSubItem}
\title{Action Subitem
Create a \code{\link[shinydashboard:sidebarMenu]{menuSubItem}}-like
\code{\link[shiny]{actionButton}} which can be passed to \code{
\link{actionItem}} as \code{...} argument.}
\usage{
actionSubItem(inputId, label)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{label}{The contents of the button or link - usually a text label, but
you could also use any other HTML, like an image.}
}
\description{
Action Subitem
Create a \code{\link[shinydashboard:sidebarMenu]{menuSubItem}}-like
\code{\link[shiny]{actionButton}} which can be passed to \code{
\link{actionItem}} as \code{...} argument.
}
