% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_utils.R
\name{append_to_reactiveValues}
\alias{append_to_reactiveValues}
\title{Append name-value pairs to a reactiveValues object}
\usage{
append_to_reactiveValues(rvs, ...)
}
\arguments{
\item{rvs}{A \code{\link[shiny]{reactiveValues}} object.}

\item{...}{\code{name = value} pairs (\code{rvs[[name]] <- value}).}
}
\description{
Create default values for a \code{\link[shiny]{reactiveValues}} object.
This function covers the special case in which a module's server function has
been called but the ui isn't part of the DOM yet, so that input values are
still missing.
}
