% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{collapsible_tabBox}
\alias{collapsible_tabBox}
\title{Create a collapsible tabbed box}
\usage{
collapsible_tabBox(..., id = NULL, selected = NULL, title = NULL,
  width = 6, height = NULL, side = c("left", "right"),
  closeable = FALSE)
}
\arguments{
\item{...}{\code{\link[shiny]{tabPanel}} elements to include in the tabset}

\item{id}{If provided, you can use \code{input$id} in your server logic to
determine which of the current tabs is active. The value will correspond to
the \code{value} argument that is passed to \code{\link[shiny]{tabPanel}}.}

\item{selected}{The \code{value} (or, if none was supplied, the \code{title})
of the tab that should be selected by default. If \code{NULL}, the first tab
will be selected.}

\item{title}{Title for the tabBox.}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default valueBox width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{height}{The height of a box, in pixels or other CSS unit. By default
the height scales automatically with the content.}

\item{side}{Which side of the box the tabs should be on (\code{"left"} or
\code{"right"}). When \code{side="right"}, the order of tabs will be
reversed.}

\item{closeable}{If the \code{\link{collapsible_tabBox}} is closeable via an
\code{\link[shiny]{actionButton}} use \code{closeable=TRUE}.}
}
\description{
Create a collapsible \code{\link[shinydashboard]{tabBox}}.
}
\details{
The \code{collapsible_tabBox} has class "collapsible-tab-box", the enclosing
div of the closing \code{\link[shiny]{actionButton}} has class
"div-btn-close".
}
