% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_deskriptive_statistik.R
\name{cumulative_distribution_function}
\alias{cumulative_distribution_function}
\title{Plot the cumulative distribution function}
\usage{
cumulative_distribution_function(tfd, breaks, col, fill, alpha, size)
}
\description{
Plot the cumulative distribution function using \code{ggplot2}.
}
