% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_utils.R
\name{firstup}
\alias{firstup}
\title{Turn the first letter of a string in upper case}
\usage{
firstup(x)
}
\arguments{
\item{x}{A character vector.}
}
\description{
Turn the first letter of a string in upper case
}
