% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{full_box}
\alias{full_box}
\title{Create a dashboard box without padding}
\usage{
full_box(..., title = NULL, footer = NULL, status = NULL,
  solidHeader = FALSE, background = NULL, width = 6, height = NULL,
  collapsible = FALSE, collapsed = FALSE)
}
\description{
Create a dashboard box without padding
}
