% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ida_transform_function_dialog.R
\name{function_dialog}
\alias{function_dialog}
\title{Initialise a dialog for executing one of the dplyr verbs on a dataset}
\usage{
function_dialog(input, output, session, .data, .values, parent,
  selected_data, dplyr_function, ...)
}
\arguments{
\item{parent}{A \code{\link{node}} object.}

\item{selected_data}{A \code{\link[shiny]{reactive}} for example returned by
\code{\link{select_data}}.}

\item{dplyr_function}{A \code{\link[shiny]{reactiveValues}} containing an
element with name 'dplyr_function' and value in \code{c("arrange", "filter",
"group_by", "mutate", "select", "summarise")}.}
}
\description{
A shiny module that calls another shiny module based on \code{dplyr_function}.
}
