% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_deskriptive_statistik.R
\name{histogram}
\alias{histogram}
\title{Plot a histogram}
\usage{
histogram(data, frequency_density = c("absolute", "relative"), breaks,
  col, fill, alpha)
}
\arguments{
\item{data}{A data frame / tibble produced for example by
\code{\link{table_frequency_distribution}}.}

\item{frequency_density}{Either "absolute" or "relative".}

\item{breaks}{A numeric vector specifying the breaks between the histogram's
groups.}

\item{col, fill, alpha}{Graphical parameters}
}
\description{
Use this function to create a simple \code{\link[ggplot2]{ggplot2}} histogram.
}
