% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_storage.R
\name{make_fluid_row}
\alias{make_fluid_row}
\title{Create a fluidRow out of a list}
\usage{
make_fluid_row(ui_fluid_row, widgets_per_row)
}
\arguments{
\item{ui_fluid_row}{A list of html fragments}

\item{widgets_per_row}{Number of widgets per row. Has to be a divisor of 12.}
}
\description{
Create a fluidRow out of a list
}
\examples{
a <- tags$div("content")
make_fluid_row(list(a, a, a, a), 4)
make_fluid_row(list(a, a, a), 4)
}
