% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{multiple_actionItem}
\alias{multiple_actionItem}
\title{Create multiple actionItem}
\usage{
multiple_actionItem(inputId_list, label_list)
}
\arguments{
\item{inputId_list}{inputIds for \code{\link[shiny]{actionButton}}. See
'Examples'.}

\item{label_list}{labels for \code{\link[shiny]{actionButton}}. See
'Examples'.}
}
\description{
Create multiple \code{\link{actionItem}}s which can in turn include
\code{\link{actionSubItem}}s.
}
\examples{

}
