% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_dashboard.R
\name{multiple_actionSubItem}
\alias{multiple_actionSubItem}
\title{Create multiple actionSubItem}
\usage{
multiple_actionSubItem(inputId_list, label_list)
}
\description{
This function is used internal in \code{\link{
multiple_actionItem}}. It could be used manually as the \code{...} argument
to \code{\link{actionItem}}.
}
