% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_utils.R
\name{multiple_menuItem}
\alias{multiple_menuItem}
\title{Call multiple menuItems in a shinydashboard sidebarMenu}
\usage{
multiple_menuItem(struct, menuItem_args = NULL,
  menuSubItem_args = NULL)
}
\arguments{
\item{menuItem_args}{Arguments passed to \code{\link[shinydashboard]{
menuItem}}.}

\item{menuSubItem_args}{Arguments passed to \code{\link[shinydashboard]{
menuSubItem}}.}

\item{struc}{Either a character vector for menuItems only or a named list
containing another named list if you want to generate menuItems with
menuSubItems. The names of the inner named list are interpreted as the
tabName whereas the content is interpreted as the text for menuSubItem.}
}
\description{
Call multiple menuItems in a shinydashboard sidebarMenu
}
