% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_mod_data_storage.R
\name{select_data}
\alias{select_data}
\title{Select data for later use in other shiny modules.}
\usage{
select_data(input, output, session, data_rvs = list(user_data_storage =
  user_data_storage, permanent_data_storage = permanent_data_storage),
  values, tabset_data = NULL, select_column = TRUE,
  grid_select = tibble(type = "select", position = 1, label =
  "Select column"), block = 2, parent, ...)
}
\arguments{
\item{values}{\code{\link[shiny]{reactiveValues}} storing data imported by the user, predefined
by the app author and other values.}

\item{tabset_data}{A tibble passed as \code{tabset_data} to
\code{\link{interact_with_tabset_panel}}.}

\item{select_column}{Deprecated. TODO: Remove}

\item{grid_select}{A \code{\link[tibble]{tibble}} specifying the displayed input
widgets for selecting the data. See 'Details'}

\item{block}{Number of input widgets displayed in a horizontal block.}

\item{parent}{A \code{\link{node}} object.}

\item{...}{Currently not used.}
}
\value{
A \code{\link[shiny]{reactiveValues}} containing information about the selection.
See 'Examples'.
}
\description{
Select data stored in one or multiple \code{\link[shiny]{reactiveValues}}.
}
\details{
\code{grid_select} has to be a \code{\link[tibble]{tibble}} with the following
columns:
\describe{
  \item{type}{A character specifing the type of the input widget. Either "select"
  for \code{\link[shiny]{selectInput}} or "selectize" for
  \code{\link[shiny]{selectizeInput}}}.
  \item{position}{A numeric. All input widgets are arranged by this in blocks
  of \code{block}.}
  \item{label}{The label for the input widget.}
}
}
\examples{

}
