% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_oo.R
\docType{data}
\name{tabBox_R6}
\alias{tabBox_R6}
\title{Create a easy accessible tabbed box}
\value{
Public methods and fields are accesible using the '$' operator.
\item{appendTab(tab, select = FALSE, closeable = TRUE)}{Append the
\code{\link[shiny]{tabPanel}} \code{tab} to the tabBox. If \code{tab} should
be selected upon being inserted use \code{select = TRUE}. If the \code{tab}
should not be closeable via an \code{\link[shiny]{actionButton}} use
\code{closeable = FALSE}.}
\item{get(what)}{Get the value of the private element with name \code{what}.}
\item{insertTab(tab, target, position = c("before", "after"),
select = FALSE, closeable = TRUE)}{Insert the \code{\link[shiny]{tabPanel}}}
\code{tab} to the tabBox. See \code{\link[shiny]{insertTab}} for Details
regarding \code{target} and \code{position}. If \code{tab} should be selected
upon being inserted use \code{select = TRUE}. If the \code{tab}
should not be closeable via an \code{\link[shiny]{actionButton}} use
\code{closeable = FALSE}.}
\item{prependTab(tab, select = FALSE, closeable = TRUE)}{Append the
\code{\link[shiny]{tabPanel}} \code{tab} to the tabBox. If \code{tab} should
be selected upon being inserted use \code{select = TRUE}. If the \code{tab}
should not be closeable via an \code{\link[shiny]{actionButton}} use
\code{closeable = FALSE}.}
\item{removeTab(target)}{Remove the tab with value \code{target}.}
\item{set_session}{Set the session for the tabBox. This method is separated
from the instantiation because you will usually want to use the tabBox
already in the ui when there is no session present yet. Therefore the session
has to be manually in the server function.
\item{tabBox()}{Return the HTML that builds the \code{\link[shinydashboard]{
tabBox}}.}
Use \code{get()} without any arguments to see the names of all private
fields and methods.
}
\description{
Create a \code{\link[shinydashboard]{tabBox}} using an R6 object. The object
contains methods for creating the tabBox, inserting and removing \code{
\link[shiny]{tabPanel}} elements and the ability to expand the usual tabBox
with an additional action button which closes a tab.
}
\details{
Instantiate a new tabBox_R6 obejct with \code{tabBox_R6$new(id,
selected = NULL, title = "Viewer", width = 6, height = NULL,
side = c("left", "right"))}.
}
\keyword{datasets}
