% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen_deskriptive_statistik.R
\name{table_frequency_distribution}
\alias{table_frequency_distribution}
\title{Table frequency distribution}
\usage{
table_frequency_distribution(x, b = NULL, k = NULL)
}
\arguments{
\item{x}{A numeric vector.}

\item{b}{The group width.}

\item{k}{The number of groups.}
}
\description{
Produce the table frequency distribution for a numeric vector.
}
\details{
Passing a value to each \code{b} and \code{k} implies grouping of the data.
}
