% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo_tab_list.R
\docType{data}
\name{tabList_R6}
\alias{tabList_R6}
\title{Create a container for ui}
\value{
Public methods and fields are accesible using the '$' operator.
\item{add_list_item_by_actionButton(ui, inputId, session)}{Add \code{ui}
(preferable_created using the method \code{tabBox}) to the \code{container}
if the \code{\link[shiny]{actionButton}} with \code{inputId=inputId} is
clicked. \code{session} is the session in which the
\code{\link[shiny]{actionButton}} is defined. This session has to be existing
at the time this method is called.}
\item{container()}{Call this method anywhere in your ui to create the
container for your elements. Note that this method can only be called once.}
\item{set_session(session)}{Set the session to a shiny session object before
you do anything with the instance in a server function.}
\item{tabBox(..., id, selected = NULL, title = NULL, width = 12,
height = NULL, side = c("left", "right"), collapsible = TRUE,
closeable = TRUE)}{Create a \code{collapsible} and/or \code{closeable}
\code{\link[shinydashboard]{tabBox}}, which has to be passed as \code{ui} to
\code{add_list_item_by_actionButton} to work properly. See
\code{\link[shinydashboard]{tabBox}} for the other arguments.}
}
\description{
Create a \code{div} that can be filled with ui from a shiny server function.
}
\details{
Instantiate a new tabList with \code{tabList_R6$new(id)}.
}
\keyword{datasets}
